space profiling done with new package data-size, test cases for prefixFilter fixed
new module Data.Size for estimating # of objects and # of words for a complex values
old statistics code put in comments
test case for prefixFilter failed. The reason
was, the result tree was not space optimized,
and there is no need to do it.
But in the test cases the expected result was space optimized,
so the expected entries matched but the structure of the trees didn't